fixes #831
