+++ horreum-backend/src/main/java/io/hyperfoil/tools/horreum/entity/data/LabelDAO.java
@@ -1,12 +1,8 @@
 package io.hyperfoil.tools.horreum.entity.data;
 
-import java.io.Serializable;
 import java.util.Collection;
-import java.util.Objects;
 
-import io.hyperfoil.tools.horreum.hibernate.JsonBinaryType;
 import jakarta.persistence.CollectionTable;
-import jakarta.persistence.Column;
 import jakarta.persistence.ElementCollection;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -15,17 +11,11 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 
 import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.annotations.Parameter;
-import org.hibernate.annotations.Type;
 import org.hibernate.id.enhanced.SequenceStyleGenerator;
-
-import com.fasterxml.jackson.databind.JsonNode;
-
-import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
 
 /* When we make changes to label we need to ensure that we remove label_values where label_id = id
 *  After delete on extractors we need to execute:
@@ -86,43 +76,4 @@ public class LabelDAO extends OwnedEntityBase {
               '}';
    }
 
-   @Entity
-   @Table(name = "label_values")
-   public static class Value extends PanacheEntityBase implements Serializable {
-      @Id
-      @NotNull
-      @Column(name = "dataset_id")
-      public int datasetId;
-
-      @Id
-      @NotNull
-      @Column(name = "label_id")
-      public int labelId;
-
-      @Type(JsonBinaryType.class)
-      @Column(columnDefinition = "jsonb")
-      public JsonNode value;
-
-      @Override
-      public boolean equals(Object o) {
-         if (this == o) return true;
-         if (o == null || getClass() != o.getClass()) return false;
-         Value value1 = (Value) o;
-         return datasetId == value1.datasetId && labelId == value1.labelId && Objects.equals(value, value1.value);
-      }
-
-      @Override
-      public int hashCode() {
-         return Objects.hash(datasetId, labelId, value);
-      }
-
-      @Override
-      public String toString() {
-         return "Value{" +
-               "datasetId=" + datasetId +
-               ", labelId=" + labelId +
-               ", value=" + value +
-               '}';
-      }
-   }
 }
